adapt artifactstore tests
